Move MinTimeTransfer to GTFSToOtpTransitServiceMapper
