+++ src/main/java/org/opentripplanner/model/OtpTransitService.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.List;
 import org.opentripplanner.ext.flex.trip.FlexTrip;
 import org.opentripplanner.model.transfer.ConstrainedTransfer;
-import org.opentripplanner.model.transfer.MinTimeTransfer;
 
 /**
  * Methods for accessing imported entities.
@@ -72,8 +71,6 @@ public interface OtpTransitService {
     List<StopTime> getStopTimesForTrip(Trip trip);
 
     Collection<ConstrainedTransfer> getAllConstrainedTransfers();
-
-    Collection<MinTimeTransfer> getMinTimeTransfers();
 
     Collection<TripPattern> getTripPatterns();
