+++ src/test/java/org/opentripplanner/graph_builder/module/ApplyMinTimeTransfersTest.java
@@ -11,7 +11,7 @@ import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.model.PathTransfer;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.plan.PlaceTest;
-import org.opentripplanner.model.transfer.MinTimeTransfer;
+import org.opentripplanner.graph_builder.model.MinTimeTransfer;
 import org.opentripplanner.routing.graph.Graph;
 
 class ApplyMinTimeTransfersTest {
